Upgrade Mirage Bower deps to latest
The app this is based on originally used an older Mirage.
The npm dependencies were updated to the newer, but not the bower.

It was hoped this would fix the issue by incorporating the fix
for https://github.com/pretenderjs/pretender/issues/85
via https://github.com/pretenderjs/pretender/pull/130
as part of the move past Pretender version 0.11.0
(https://github.com/pretenderjs/pretender/blob/master/CHANGELOG.md#0110),
but, surprisingly, that did *not* fix the issue.

Well, this is frustrating.