Fix minimal example by altering response type
Exceptions appear to be swallowed by test environment.
Setting an appropriate breakpoint and trapping on all exceptions,
including caught ones, shows that it was trying to parse the string
as JSON. Returning an object allowed things to proceed.

The original app where this problem occurs is actually using
passthrough, without any handlers, so this means that I failed to
minimize properly in converting from the passthrough to the
non-passthrough interceptor.

The exception trapping behavior though suggests I must
investigate possible silent exceptions triggering the problem.